Controllers, Validators y Swagger

+++ SolvexApi.Services/Services/BaseService.cs
@@ -1,5 +1,8 @@
 ﻿using AutoMapper;
+using FluentValidation;
 using Microsoft.EntityFrameworkCore;
+using SolvexApi.Bl.Extensions;
+using SolvexApi.Core.Adstract;
 using SolvexApi.Core.BaseModel.Base;
 using SolvexApi.Model.Interfaces;
 using System;
@@ -13,9 +16,9 @@ namespace SolvexApi.Services.Services
     {
         Task<IEnumerable<TDto>> GetAllAsync();
         Task<TDto> GetAsync(int id);
-        Task<TDto> AddAsync(TDto dto);
-        Task<TDto> UpdateAsync(int id, TDto dto);
-        Task<TDto> DeleteByIdAsync(int id);
+        Task<IEntityOperationResult<TDto>> AddAsync(TDto dto);
+        Task<IEntityOperationResult<TDto>> UpdateAsync(int id, TDto dto);
+        Task<IEntityOperationResult<TDto>> DeleteByIdAsync(int id);
     }
     public class BaseService<TEntity, TDto> : IBaseService<TEntity, TDto>
         where TEntity : class, IBase
@@ -23,11 +26,13 @@ namespace SolvexApi.Services.Services
     {
         protected readonly IBaseRepository<TEntity> _repository;
         protected readonly IMapper _mapper;
+        protected readonly IValidator<TDto> _validator;
 
-        public BaseService(IBaseRepository<TEntity> repository, IMapper mapper)
+        public BaseService(IBaseRepository<TEntity> repository, IMapper mapper, IValidator<TDto> validator)
         {
             _repository = repository;
             _mapper = mapper;
+            _validator = validator;
         }
 
         public async Task<IEnumerable<TDto>> GetAllAsync()
@@ -44,29 +49,53 @@ namespace SolvexApi.Services.Services
             return dto;
         }
 
-        public async Task<TDto> AddAsync(TDto dto)
+        public async Task<IEntityOperationResult<TDto>> AddAsync(TDto dto)
         {
+            var validationResult = _validator.Validate(dto);
+            if(validationResult.IsValid is false)
+                return validationResult.ToOperationResult<TDto>();
+
             TEntity entity = _mapper.Map<TEntity>(dto);
-            var result = await _repository.Add(entity);
-            _mapper.Map(result, dto);
-            return dto;
+            var Entityresult = await _repository.Add(entity);
+
+            _mapper.Map(Entityresult, dto);
+
+            var result = dto.ToOperationResult<TDto>();
+            return result;
         }
 
-        public async Task<TDto> UpdateAsync(int id, TDto dto)
+        public async Task<IEntityOperationResult<TDto>> UpdateAsync(int id, TDto dto)
         {
+            var validationResult = _validator.Validate(dto);
+            if (validationResult.IsValid is false)
+                return validationResult.ToOperationResult<TDto>();
+
             var entity = await _repository.Get(id);
+
+            if (entity is null)
+                return null;
+
             _mapper.Map(dto, entity);
 
-            var result = await _repository.Update(entity);
-            _mapper.Map(result, dto);
-            return dto;
+             entity = await _repository.Update(entity);
+            _mapper.Map(entity, dto);
+
+            var result = dto.ToOperationResult<TDto>();
+            return result;
         }
 
-        public async Task<TDto> DeleteByIdAsync(int id)
+        public async Task<IEntityOperationResult<TDto>> DeleteByIdAsync(int id)
         {
-            var entity = await _repository.Delete(id);
+            var entity = await _repository.Get(id);
+
+            if (entity is null)
+                return null;
+
+            entity = await _repository.Delete(id);
             var dto = _mapper.Map<TDto>(entity);
-            return dto;
+
+            var result = dto.ToOperationResult<TDto>();
+            return result;
         }
     }
 }
